apps/backend/gateway/requirements.txt
# API Gateway Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Authentication & Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyotp==2.9.0
qrcode[pil]==7.4.2

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Redis
redis==5.0.1
hiredis==2.2.3

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pydantic-settings==2.1.0

# Database Migrations
alembic==1.12.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
